Organize client and service into package
